Initial work for implementing stoppable harvesting sessions

+++ geonode/harvesting/tasks.py
@@ -28,13 +28,10 @@ from django.db.models import (
     Value,
 )
 from django.db.models.functions import Concat
-from django.utils.timezone import now
+from django.utils import timezone
 from geonode.celery_app import app
 
-from . import (
-    models,
-    utils,
-)
+from . import models
 from .harvesters import base
 
 logger = logging.getLogger(__name__)
@@ -48,7 +45,6 @@ logger = logging.getLogger(__name__)
 )
 def harvesting_dispatcher(
         self,
-        harvester_id: int,
         harvesting_session_id: int
 ):
     """Perform harvesting asynchronously.
@@ -71,7 +67,7 @@ def harvesting_dispatcher(
 
     """
 
-    harvester = models.Harvester.objects.get(pk=harvester_id)
+    harvester = models.Harvester.objects.get(harvesting_session__pk=harvesting_session_id)
     harvestable_resources = list(harvester.harvestable_resources.filter(
         should_be_harvested=True).values_list("id", flat=True))
     if len(harvestable_resources) > 0:
@@ -117,8 +113,7 @@ def harvest_resources(
             )
         else:
             harvester = sample_harvestable_resource.harvester
-            available = utils.update_harvester_availability(harvester)
-            if available:
+            if harvester.update_availability():
                 harvester.status = harvester.STATUS_PERFORMING_HARVESTING
                 harvester.save()
                 models.HarvestingSession.objects.filter(pk=harvesting_session_id).update(
@@ -172,7 +167,7 @@ def _harvest_resource(
     harvestable_resource = models.HarvestableResource.objects.get(pk=harvestable_resource_id)
     worker: base.BaseHarvesterWorker = harvestable_resource.harvester.get_harvester_worker()
     harvested_resource_info = worker.get_resource(harvestable_resource, harvesting_session_id)
-    now_ = now()
+    now_ = timezone.now()
     if harvested_resource_info is not None:
         if worker.should_copy_resource(harvestable_resource):
             copied_path = worker.copy_resource(harvestable_resource, harvested_resource_info)
@@ -276,7 +271,7 @@ def _handle_harvesting_error(self, task_id, *args, **kwargs):
 )
 def check_harvester_available(self, harvester_id: int):
     harvester = models.Harvester.objects.get(pk=harvester_id)
-    available = utils.update_harvester_availability(harvester)
+    available = harvester.update_availability()
     logger.info(
         f"Harvester {harvester!r}: remote server is "
         f"{'' if available else 'not '}available"
@@ -289,42 +284,55 @@ def check_harvester_available(self, harvester_id: int):
     acks_late=False,
     ignore_result=False,
 )
-def update_harvestable_resources(self, harvester_id: int):
+def update_harvestable_resources(self, refresh_session_id: int):
     # NOTE: we are able to implement batch discovery of existing harvestable resources
     # because we want to know about all of them. We are not able to batch harvesting
     # of resources because these have potentially been individually selected by the
     # user, which means we are not interested in all of them
-    harvester = models.Harvester.objects.get(pk=harvester_id)
-    harvester.status = harvester.STATUS_UPDATING_HARVESTABLE_RESOURCES
-    harvester.save()
-    worker = harvester.get_harvester_worker()
-    try:
-        num_resources = worker.get_num_available_resources()
-    except (NotImplementedError, base.HarvestingException) as exc:
-        _handle_harvestable_resources_update_error(
-            self.request.id, harvester_id=harvester_id, raised_exception=exc)
-    else:
-        harvester.num_harvestable_resources = num_resources
-        harvester.save()
-        page_size = 10
-        total_pages = math.ceil(num_resources / page_size)
-        batches = []
-        for page in range(total_pages):
-            batches.append(
-                _update_harvestable_resources_batch.signature(
-                    args=(harvester_id, page, page_size),
+    refresh_session = models.HarvestableResourceRefreshSession.objects.get(pk=refresh_session_id)
+    if refresh_session.status != refresh_session.STATUS_ABORTED:
+        refresh_session.status = models.HarvestableResourceRefreshSession.STATUS_ON_GOING
+        refresh_session.save()
+        harvester = refresh_session.harvester
+        if harvester.update_availability():
+            harvester.status = harvester.STATUS_UPDATING_HARVESTABLE_RESOURCES
+            harvester.save()
+            worker = harvester.get_harvester_worker()
+            try:
+                num_resources = worker.get_num_available_resources()
+            except (NotImplementedError, base.HarvestingException) as exc:
+                _handle_harvestable_resources_update_error(
+                    self.request.id, refresh_session_id=refresh_session_id, raised_exception=exc)
+            else:
+                harvester.num_harvestable_resources = num_resources
+                harvester.save()
+                refresh_session.remote_records = num_resources
+                refresh_session.save()
+                page_size = 10
+                total_pages = math.ceil(num_resources / page_size)
+                batches = []
+                for page in range(total_pages):
+                    batches.append(
+                        _update_harvestable_resources_batch.signature(
+                            args=(refresh_session_id, page, page_size),
+                        )
+                    )
+                update_finalizer = _finish_harvestable_resources_update.signature(
+                    args=(refresh_session_id,),
+                    immutable=True
+                ).on_error(
+                    _handle_harvestable_resources_update_error.signature(
+                        kwargs={"refresh_session_id": refresh_session_id}
+                    )
                 )
+                update_workflow = chord(batches, body=update_finalizer)
+                update_workflow.apply_async()
+        else:
+            finish_refresh_session(
+                refresh_session_id,
+                models.HarvestableResourceRefreshSession.STATUS_FINISHED_ALL_FAILED,
+                final_details="Harvester is not available"
             )
-        update_finalizer = _finish_harvestable_resources_update.signature(
-            args=(harvester_id,),
-            immutable=True
-        ).on_error(
-            _handle_harvestable_resources_update_error.signature(
-                kwargs={"harvester_id": harvester_id}
-            )
-        )
-        update_workflow = chord(batches, body=update_finalizer)
-        update_workflow.apply_async()
 
 
 @app.task(
@@ -334,31 +342,35 @@ def update_harvestable_resources(self, harvester_id: int):
     ignore_result=False,
 )
 def _update_harvestable_resources_batch(
-        self, harvester_id: int, page: int, page_size: int):
-    harvester = models.Harvester.objects.get(pk=harvester_id)
-    worker = harvester.get_harvester_worker()
-    offset = page * page_size
-    try:
-        found_resources = worker.list_resources(offset)
-    except base.HarvestingException:
-        logger.exception("Could not retrieve list of remote resources.")
+        self, refresh_session_id: int, page: int, page_size: int):
+    refresh_session = models.HarvestableResourceRefreshSession.objects.get(pk=refresh_session_id)
+    if refresh_session.status == refresh_session.STATUS_ON_GOING:
+        harvester = refresh_session.harvester
+        worker = harvester.get_harvester_worker()
+        offset = page * page_size
+        try:
+            found_resources = worker.list_resources(offset)
+        except base.HarvestingException:
+            logger.exception("Could not retrieve list of remote resources.")
+        else:
+            for remote_resource in found_resources:
+                resource, created = models.HarvestableResource.objects.get_or_create(
+                    harvester=harvester,
+                    unique_identifier=remote_resource.unique_identifier,
+                    title=remote_resource.title,
+                    defaults={
+                        "should_be_harvested": harvester.harvest_new_resources_by_default,
+                        "remote_resource_type": remote_resource.resource_type,
+                        "last_refreshed": timezone.now()
+                    }
+                )
+                # NOTE: make sure to save the resource because we need to have its
+                # `last_updated` property be refreshed - this is done in order to be able
+                # to compare when a resource has been found
+                resource.last_refreshed = timezone.now()
+                resource.save()
     else:
-        for remote_resource in found_resources:
-            resource, created = models.HarvestableResource.objects.get_or_create(
-                harvester=harvester,
-                unique_identifier=remote_resource.unique_identifier,
-                title=remote_resource.title,
-                defaults={
-                    "should_be_harvested": harvester.harvest_new_resources_by_default,
-                    "remote_resource_type": remote_resource.resource_type,
-                    "last_refreshed": now()
-                }
-            )
-            # NOTE: make sure to save the resource because we need to have its
-            # `last_updated` property be refreshed - this is done in order to be able
-            # to compare when a resource has been found
-            resource.last_refreshed = now()
-            resource.save()
+        logger.info(f"The refresh session has been asked to abort, so skipping...")
 
 
 @app.task(
@@ -367,15 +379,24 @@ def _update_harvestable_resources_batch(
     acks_late=False,
     ignore_result=False,
 )
-def _finish_harvestable_resources_update(self, harvester_id: int):
-    harvester = models.Harvester.objects.get(pk=harvester_id)
-    if harvester.last_checked_harvestable_resources is not None:
-        _delete_stale_harvestable_resources(harvester)
+def _finish_harvestable_resources_update(self, refresh_session_id: int):
+    refresh_session = models.HarvestableResourceRefreshSession.objects.get(pk=refresh_session_id)
+    harvester = refresh_session.harvester
+    if refresh_session.status == refresh_session.STATUS_ABORTING:
+        message = "Refresh session aborted by user"
+        finish_refresh_session(
+            refresh_session_id, refresh_session.STATUS_ABORTED, final_details=message)
+    else:
+        message = "Harvestable resources successfully checked"
+        finish_refresh_session(
+            refresh_session_id, refresh_session.STATUS_FINISHED_ALL_OK, final_details=message)
+        if harvester.last_checked_harvestable_resources is not None:
+            _delete_stale_harvestable_resources(harvester)
+    refresh_session.save()
     harvester.status = harvester.STATUS_READY
-    now_ = now()
+    now_ = timezone.now()
     harvester.last_checked_harvestable_resources = now_
-    harvester.last_check_harvestable_resources_message = (
-        f"{now_} - Harvestable resources successfully checked")
+    harvester.last_check_harvestable_resources_message = f"{now_} - {message}"
     harvester.save()
 
 
@@ -386,14 +407,17 @@ def _finish_harvestable_resources_update(self, harvester_id: int):
     ignore_result=False,
 )
 def _handle_harvestable_resources_update_error(self, task_id, *args, **kwargs):
+    logger.debug("Inside _handle_harvestable_resources_update_error -----------------------------------------")
     result = self.app.AsyncResult(str(task_id))
-    print(f"locals: {locals()}")
-    print(f"state: {result.state}")
-    print(f"result: {result.result}")
-    print(f"traceback: {result.traceback}")
-    harvester = models.Harvester.objects.get(pk=kwargs["harvester_id"])
+    logger.debug(f"locals: {locals()}")
+    logger.debug(f"state: {result.state}")
+    logger.debug(f"result: {result.result}")
+    logger.debug(f"traceback: {result.traceback}")
+    result = self.app.AsyncResult(str(task_id))
+    refresh_session = models.HarvestableResourceRefreshSession.objects.get(pk=kwargs["refresh_session_id"])
+    harvester = refresh_session.harvester
     harvester.status = harvester.STATUS_READY
-    now_ = now()
+    now_ = timezone.now()
     harvester.last_checked_harvestable_resources = now_
     harvester.last_check_harvestable_resources_message = (
         f"{now_} - There was an error retrieving information on available "
@@ -401,6 +425,12 @@ def _handle_harvestable_resources_update_error(self, task_id, *args, **kwargs):
         f"Please check the logs"
     )
     harvester.save()
+    details = f"state: {result.state}\nresult: {result.result}\ntraceback: {result.traceback}"
+    finish_refresh_session(
+        kwargs["refresh_session_id"],
+        final_status=models.HarvestableResourceRefreshSession.STATUS_FINISHED_SOME_FAILED,
+        final_details=details
+    )
 
 
 def _delete_stale_harvestable_resources(harvester: models.Harvester):
@@ -422,7 +452,7 @@ def _delete_stale_harvestable_resources(harvester: models.Harvester):
 
     previously_checked_at = harvester.last_checked_harvestable_resources
     logger.debug(f"last checked at: {previously_checked_at}")
-    logger.debug(f"now: {now()}")
+    logger.debug(f"now: {timezone.now()}")
     to_remove = models.HarvestableResource.objects.filter(
         harvester=harvester, last_refreshed__lte=previously_checked_at)
     for harvestable_resource in to_remove:
@@ -461,7 +491,7 @@ def finish_harvesting_session(
 ) -> None:
     """Finish the input harvesting session"""
     update_kwargs = {
-        "ended": now(),
+        "ended": timezone.now(),
         "status": final_status,
     }
     if additional_harvested_records is not None:
@@ -470,3 +500,40 @@ def finish_harvesting_session(
     if final_details is not None:
         update_kwargs["session_details"] = Concat("session_details", Value(f"\n{final_details}"))
     models.HarvestingSession.objects.filter(id=session_id).update(**update_kwargs)
+
+
+def update_refresh_session(
+        refresh_session_id: int,
+        total_records_found: typing.Optional[int] = None,
+        additional_harvested_records: typing.Optional[int] = None,
+        additional_details: typing.Optional[str] = None,
+) -> None:
+    """Update the input harvesting session."""
+    update_kwargs = {}
+    if total_records_found is not None:
+        update_kwargs["total_records_found"] = total_records_found
+    if additional_harvested_records is not None:
+        update_kwargs["records_harvested"] = (
+                F("records_harvested") + additional_harvested_records)
+    if additional_details is not None:
+        update_kwargs["session_details"] = Concat("session_details", Value(f"\n{additional_details}"))
+    models.HarvestableResourceRefreshSession.objects.filter(id=refresh_session_id).update(**update_kwargs)
+
+
+def finish_refresh_session(
+        refresh_session_id: int,
+        final_status: str,
+        final_details: typing.Optional[str] = None,
+        additional_harvested_records: typing.Optional[int] = None,
+) -> None:
+    """Finish the input refresh session"""
+    update_kwargs = {
+        "ended": timezone.now(),
+        "status": final_status,
+    }
+    if additional_harvested_records is not None:
+        update_kwargs["records_harvested"] = (
+                F("records_harvested") + additional_harvested_records)
+    if final_details is not None:
+        update_kwargs["session_details"] = Concat("session_details", Value(f"\n{final_details}"))
+    models.HarvestableResourceRefreshSession.objects.filter(id=refresh_session_id).update(**update_kwargs)
